Switch hook to function to avoid dependency needs
