+++ src/views/v2/Bridge/Routes/SingleRoute.tsx
@@ -31,7 +31,7 @@ import { useGetTokens } from 'hooks/useGetTokens';
 import { useTokens } from 'contexts/TokensContext';
 import GasSlider from 'views/v2/Bridge/GasSlider';
 import Color from 'color';
-import { useNumberFormatter } from 'hooks/useNumberFormatter';
+import { formatWithCommas } from 'utils/formatNumber';
 
 const HIGH_FEE_THRESHOLD = 20; // dollhairs
 
@@ -49,7 +49,6 @@ type Props = {
 
 const SingleRoute = (props: Props) => {
   const theme = useTheme();
-  const { formatWithCommas } = useNumberFormatter();
   const styles = useMemo(
     () => ({
       container: {
@@ -421,12 +420,14 @@ const SingleRoute = (props: Props) => {
 
   const receiveAmountTrunc = useMemo(() => {
     if (quote) {
-      const truncatedAmount = amount.display(amount.truncate(quote.destinationToken.amount, 6));
+      const truncatedAmount = amount.display(
+        amount.truncate(quote.destinationToken.amount, 6),
+      );
       return formatWithCommas(truncatedAmount);
     } else {
       return undefined;
     }
-  }, [quote, formatWithCommas]);
+  }, [quote]);
 
   const routeCardHeader = useMemo(() => {
     if (props.error) {
@@ -492,7 +493,6 @@ const SingleRoute = (props: Props) => {
     providerText,
     receiveAmount,
     theme.palette.text.secondary,
-    formatWithCommas,
   ]);
 
   // There are three states for the Card area cursor:
