+++ src/views/v3/Bridge/AmountInput/index.tsx
@@ -19,7 +19,7 @@ import Box from '@mui/material/Box';
 import { Token } from 'config/tokens';
 import type { RootState } from 'store';
 import { useGetTokens } from 'hooks/useGetTokens';
-import { useNumberFormatter } from 'hooks/useNumberFormatter';
+import { formatWithCommas, removeCommas } from 'utils/formatNumber';
 
 const INPUT_DEBOUNCE = 500;
 
@@ -36,7 +36,6 @@ const DebouncedTextField = memo(
   }) => {
     const [innerValue, setInnerValue] = useState<string>(value ?? '');
     const [isFocused, setIsFocused] = useState(false);
-    const { formatWithCommas, removeCommas } = useNumberFormatter();
     const deferredOnChange = useDebouncedCallback(
       onDebouncedChange,
       INPUT_DEBOUNCE,
@@ -68,7 +67,7 @@ const DebouncedTextField = memo(
         onChange(valueWithoutFormatting); // callback with no delay
         deferredOnChange(valueWithoutFormatting);
       },
-      [deferredOnChange, onChange, formatWithCommas, removeCommas],
+      [deferredOnChange, onChange],
     );
 
     // Propagate any outside changes to the inner TextField value
@@ -80,7 +79,7 @@ const DebouncedTextField = memo(
       }
       // We should run this side-effect only when the value changes
       // eslint-disable-next-line react-hooks/exhaustive-deps
-    }, [value, formatWithCommas]);
+    }, [value]);
 
     return (
       <TextField
